+++ common/src/main/java/com/github/yajatkaul/mega_showdown/mixin/battle/BattleTargetSelectionMixin.java
@@ -16,7 +16,8 @@ public abstract class BattleTargetSelectionMixin {
     @Shadow
     private String gimmickID;
     @Final
-    @Shadow private InBattleMove move;
+    @Shadow
+    private InBattleMove move;
 
     @Redirect(
             method = "<init>",