+++ common/src/main/java/com/github/yajatkaul/mega_showdown/mixin/battle/MoveActionResponseMixin.java
@@ -1,13 +1,12 @@
 package com.github.yajatkaul.mega_showdown.mixin.battle;
 
-import com.cobblemon.mod.common.api.battles.model.actor.BattleActor;
-import com.cobblemon.mod.common.battles.*;
-import kotlin.Pair;
+import com.cobblemon.mod.common.battles.ActiveBattlePokemon;
+import com.cobblemon.mod.common.battles.MoveActionResponse;
+import com.cobblemon.mod.common.battles.ShowdownMoveset;
+import com.github.yajatkaul.mega_showdown.cobblemon.battle.MoveActionResponseMSD;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Overwrite;
 import org.spongepowered.asm.mixin.Shadow;
-
-import java.util.List;
 
 @Mixin(value = MoveActionResponse.class, remap = false)
 public class MoveActionResponseMixin {
@@ -19,37 +18,11 @@ public class MoveActionResponseMixin {
     private String gimmickID;
 
     /**
-     * @author YajatKaul, Provismet
+     * @author YajatKaul
      * @reason TargetSelection
      */
     @Overwrite
     public boolean isValid(ActiveBattlePokemon activeBattlePokemon, ShowdownMoveset showdownMoveSet, boolean forceSwitch) {
-        if (forceSwitch || showdownMoveSet == null) {
-            return false;
-        }
-
-        InBattleMove move = showdownMoveSet.getMoves().stream()
-                .filter(m -> m.getId().equals(moveName))
-                .findFirst()
-                .orElse(null);
-        if (move == null) return false;
-
-        InBattleGimmickMove gimmickMove = move.getGimmickMove();
-        boolean validGimmickMove = gimmickMove != null && !gimmickMove.getDisabled();
-        if (!validGimmickMove && !move.canBeUsed()) {
-            return false;
-        }
-
-        List<Targetable> availableTargets;
-        if (gimmickID != null && validGimmickMove && !gimmickID.equalsIgnoreCase("mega") && !gimmickID.equalsIgnoreCase("terastal")) {
-            availableTargets = gimmickMove.getTarget().getTargetList().invoke(activeBattlePokemon);
-        } else {
-            availableTargets = move.getTarget().getTargetList().invoke(activeBattlePokemon);
-        }
-
-        if (availableTargets == null || availableTargets.isEmpty()) return true;
-        if (this.targetPnx == null) return false;
-        Pair<BattleActor, ActiveBattlePokemon> targetPair = activeBattlePokemon.getActor().getBattle().getActorAndActiveSlotFromPNX(this.targetPnx);
-        return availableTargets.contains(targetPair.getSecond());
+        return MoveActionResponseMSD.isValid(activeBattlePokemon, showdownMoveSet, forceSwitch, moveName, targetPnx, gimmickID);
     }
 }