+++ common/src/main/java/com/github/yajatkaul/mega_showdown/mixin/client/PokemonRendererMixin.java
@@ -33,7 +33,10 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import java.util.*;
+import java.util.HashSet;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
 
 /*
 IK this code is shit but im bored doing this so ill fix it later