+++ common/src/main/java/com/github/yajatkaul/mega_showdown/render/layerEntities/DynamaxCloudsLayer.java
@@ -20,7 +20,6 @@ import net.minecraft.client.renderer.texture.OverlayTexture;
 import net.minecraft.resources.ResourceLocation;
 
 import java.util.HashSet;
-import java.util.List;
 import java.util.Map;
 import java.util.Set;