+++ common/src/main/java/com/github/yajatkaul/mega_showdown/render/layerEntities/TeraHatsLayer.java
@@ -21,7 +21,6 @@ import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.texture.OverlayTexture;
 import net.minecraft.resources.ResourceLocation;
 
-import java.util.List;
 import java.util.Map;
 import java.util.Set;