+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # 1.6.13+1.7.3-release
 - Fixed interaction wheel gimmick usage during battle
 - Fixed du fusions bug where you could use it on two fusion mons one by one and it would remove the previous one stored
+- "Allegedly" fixes battle freezes
+- Fixed custom types crashing in multiplayer